Move login dialog module to modules root
